split services and clients, get tokens from controller
