+++ modules/travel_pay/spec/services/claims_client_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe TravelPay::Client do
+describe TravelPay::ClaimsClient do
   let(:user) { build(:user) }
 
   before do
@@ -14,7 +14,7 @@ describe TravelPay::Client do
       c.request :json
     end
 
-    allow_any_instance_of(TravelPay::Client).to receive(:connection).and_return(conn)
+    allow_any_instance_of(TravelPay::ClaimsClient).to receive(:connection).and_return(conn)
   end
 
   context 'prod settings' do
@@ -77,15 +77,15 @@ describe TravelPay::Client do
   context '/claims' do
     before do
       allow_any_instance_of(TravelPay::TokenService)
-        .to receive(:request_veis_token)
-        .and_return('veis_token')
-      # allow_any_instance_of(TravelPay::TokenService)
-      #   .to receive(:request_sts_token)
-      #   .and_return('sts_token')
-      allow_any_instance_of(TravelPay::TokenService)
-        .to receive(:request_btsss_token)
-        .with(user)
-        .and_return('btsss_token')
+        .to receive(:get_tokens)
+        .and_return({ 'veis_token' => 'veis_token', 'btsss_token' => 'btsss_token' })
+      #   # allow_any_instance_of(TravelPay::TokenService)
+      #   #   .to receive(:request_sts_token)
+      #   #   .and_return('sts_token')
+      #   # allow_any_instance_of(TravelPay::TokenService)
+      #   #   .to receive(:request_btsss_token)
+      #   #   .with(user)
+      #   #   .and_return('btsss_token')
     end
 
     it 'returns response from claims endpoint' do
@@ -129,52 +129,11 @@ describe TravelPay::Client do
 
       expected_ids = %w[uuid1 uuid2 uuid3]
 
-      client = TravelPay::Client.new
-      claims_response = client.get_claims(user)
+      client = TravelPay::ClaimsClient.new
+      claims_response = client.get_claims('veis_token', 'btsss_token')
       actual_claim_ids = claims_response.body['data'].pluck('id')
 
       expect(actual_claim_ids).to eq(expected_ids)
     end
   end
-
-  # context 'request_sts_token' do
-  #   let(:assertion) do
-  #     {
-  #       'iss' => 'https://www.example.com',
-  #       'sub' => user.email,
-  #       'aud' => 'https://www.example.com/v0/sign_in/token',
-  #       'iat' => 1_634_745_556,
-  #       'exp' => 1_634_745_856,
-  #       'scopes' => [],
-  #       'service_account_id' => nil,
-  #       'jti' => 'c3fa0763-70cb-419a-b3a6-d2563e7b8504',
-  #       'user_attributes' => { 'icn' => '123498767V234859' }
-  #     }
-  #   end
-  #   let(:grant_type) { 'urn:ietf:params:oauth:grant-type:jwt-bearer' }
-
-  #   before do
-  #     Timecop.freeze(Time.zone.parse('2021-10-20T15:59:16Z'))
-  #     allow(SecureRandom).to receive(:uuid).and_return('c3fa0763-70cb-419a-b3a6-d2563e7b8504')
-  #   end
-
-  #   after { Timecop.return }
-
-  #   it 'builds sts assertion and requests sts token' do
-  #     private_key_file = Settings.sign_in.sts_client.key_path
-  #     private_key = OpenSSL::PKey::RSA.new(File.read(private_key_file))
-  #     jwt = JWT.encode(assertion, private_key, 'RS256')
-  #     @stubs.post("http:/v0/sign_in/token?assertion=#{jwt}&grant_type=#{grant_type}") do
-  #       [
-  #         200,
-  #         { 'Content-Type': 'application/json' },
-  #         '{"data": {"access_token": "fake_sts_token"}}'
-  #       ]
-  #     end
-  #     client = TravelPay::Client.new
-  #     sts_token = client.request_sts_token(user)
-  #     expect(sts_token).to eq('fake_sts_token')
-  #     @stubs.verify_stubbed_calls
-  #   end
-  # end
 end
